+++ src/core/Table/Table.test.tsx
@@ -13,7 +13,7 @@ import React from 'react';
 import { Table, TableProps } from './Table';
 import * as IntersectionHooks from '../utils/hooks/useIntersection';
 import { tableFilters } from './filters';
-import { CellProps } from 'react-table';
+import { CellProps, Row } from 'react-table';
 import { SvgChevronRight } from '@itwin/itwinui-icons-react';
 
 const intersectionCallbacks = new Map<Element, () => void>();
@@ -969,4 +969,23 @@ it('should select and filter rows', () => {
   expect(checkboxInputs[1].checked).toBe(true);
   expect(checkboxInputs[2].checked).toBe(true);
   expect(checkboxInputs[3].checked).toBe(false);
+});
+
+it('should pass custom props to row', () => {
+  const onMouseEnter = jest.fn();
+  let element: HTMLInputElement | null = null;
+  const onRef = (ref: HTMLInputElement) => {
+    element = ref;
+  };
+  const rowProps = (row: Row<{ name: string; description: string }>) => {
+    return { onMouseEnter: () => onMouseEnter(row.original), ref: onRef };
+  };
+  const { container } = renderComponent({ rowProps });
+
+  const rows = container.querySelectorAll('.iui-table-body .iui-row');
+  expect(rows.length).toBe(3);
+
+  fireEvent.mouseEnter(rows[0]);
+  expect(onMouseEnter).toHaveBeenCalledWith(mockedData()[0]);
+  expect(element).toBeTruthy();
 });