+++ src/core/Table/Table.tsx
@@ -129,6 +129,11 @@ export type TableProps<
    * If not specified, all rows are enabled.
    */
   isRowDisabled?: (rowData: T) => boolean;
+  /**
+   * Function that should return custom props passed to the each row.
+   * Must be memoized.
+   */
+  rowProps?: (row: Row<T>) => React.ComponentPropsWithRef<'div'>;
 } & Omit<CommonProps, 'title'>;
 
 /**
@@ -201,6 +206,7 @@ export const Table = <
     filterTypes: filterFunctions,
     expanderCell,
     isRowDisabled,
+    rowProps,
     ...rest
   } = props;
 
@@ -546,13 +552,6 @@ export const Table = <
         {data.length !== 0 &&
           rows.map((row: Row<T>) => {
             prepareRow(row);
-            const rowProps = row.getRowProps({
-              className: cx('iui-row', {
-                'iui-selected': row.isSelected,
-                'iui-row-expanded': row.isExpanded && subComponent,
-                'iui-disabled': isRowDisabled?.(row.original),
-              }),
-            });
             return (
               <TableRowMemoized
                 row={row}
@@ -562,9 +561,10 @@ export const Table = <
                 onBottomReached={onBottomReachedRef}
                 intersectionMargin={intersectionMargin}
                 state={state}
-                key={rowProps.key}
+                key={row.getRowProps().key}
                 onClick={onRowClickHandler}
                 subComponent={subComponent}
+                isDisabled={!!isRowDisabled?.(row.original)}
               />
             );
           })}