+++ src/core/utils/hooks/useMergedRefs.ts
@@ -7,7 +7,9 @@ import React from 'react';
 /**
  * Function that merges the provided refs into one.
  */
-export const mergeRefs = <T>(...refs: Array<React.Ref<T>>) => {
+export const mergeRefs = <T>(
+  ...refs: Array<React.Ref<T> | undefined | null>
+) => {
   return (instance: T | null) => {
     refs.forEach((ref) => {
       if (typeof ref === 'function') {
@@ -22,7 +24,9 @@ export const mergeRefs = <T>(...refs: Array<React.Ref<T>>) => {
 /**
  * Returns a ref callback that merges the provided refs using `mergeRefs`.
  */
-export const useMergedRefs = <T>(...refs: ReadonlyArray<React.Ref<T>>) => {
+export const useMergedRefs = <T>(
+  ...refs: ReadonlyArray<React.Ref<T> | undefined | null>
+) => {
   // eslint-disable-next-line react-hooks/exhaustive-deps
   return React.useCallback(mergeRefs(...refs), [...refs]);
 };